SW-2013, SW-2014: split update unit test
